Enhance README with more detailed package description
Added a more comprehensive and detailed description of databot-py in README. This includes an explanation of how the Python interface interacts with the databot sensor device, the features of the PyDatabot class, and potential ways of retrieving sensor values. This should provide potential users with a more thorough understanding of the package functions and capabilities.